Replaced Infactor URL with a working one.
The Infactor.io URL linked does not work. Checking their historical data, the link was
infactor.com however this website is also no longer online. A working link we can use
is from the Wayback Machine and uses the if_ flag in the URL to remove the Web Archive
toolbar.