Changed Kramaa website link to a working one.
Kramaa.com is no longer online so the link does not work. The Wayback Machine
still has an archived copy. If we load this link using the if_ flag in the
URL, we can load the page without the Wayback Machine toolbar. This commit
replaces the kramaa.com link with the Wayback machine link as described.